pyproject.toml
[project]
name = "agentkred"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.3",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "greenlet>=3.3.1",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pynacl>=1.6.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
